+++ capiscio_sdk/_rpc/gen/capiscio/v1/__init__.py
@@ -1,16 +1,16 @@
 # Capiscio v1 protobuf modules
-from capiscio_sdk._rpc.gen.capiscio.v1 import badge_pb2
-from capiscio_sdk._rpc.gen.capiscio.v1 import badge_pb2_grpc
-from capiscio_sdk._rpc.gen.capiscio.v1 import common_pb2
-from capiscio_sdk._rpc.gen.capiscio.v1 import did_pb2
-from capiscio_sdk._rpc.gen.capiscio.v1 import did_pb2_grpc
-from capiscio_sdk._rpc.gen.capiscio.v1 import registry_pb2
-from capiscio_sdk._rpc.gen.capiscio.v1 import registry_pb2_grpc
-from capiscio_sdk._rpc.gen.capiscio.v1 import revocation_pb2
-from capiscio_sdk._rpc.gen.capiscio.v1 import revocation_pb2_grpc
-from capiscio_sdk._rpc.gen.capiscio.v1 import scoring_pb2
-from capiscio_sdk._rpc.gen.capiscio.v1 import scoring_pb2_grpc
-from capiscio_sdk._rpc.gen.capiscio.v1 import simpleguard_pb2
-from capiscio_sdk._rpc.gen.capiscio.v1 import simpleguard_pb2_grpc
-from capiscio_sdk._rpc.gen.capiscio.v1 import trust_pb2
-from capiscio_sdk._rpc.gen.capiscio.v1 import trust_pb2_grpc
+from capiscio_sdk._rpc.gen.capiscio.v1 import badge_pb2 as badge_pb2
+from capiscio_sdk._rpc.gen.capiscio.v1 import badge_pb2_grpc as badge_pb2_grpc
+from capiscio_sdk._rpc.gen.capiscio.v1 import common_pb2 as common_pb2
+from capiscio_sdk._rpc.gen.capiscio.v1 import did_pb2 as did_pb2
+from capiscio_sdk._rpc.gen.capiscio.v1 import did_pb2_grpc as did_pb2_grpc
+from capiscio_sdk._rpc.gen.capiscio.v1 import registry_pb2 as registry_pb2
+from capiscio_sdk._rpc.gen.capiscio.v1 import registry_pb2_grpc as registry_pb2_grpc
+from capiscio_sdk._rpc.gen.capiscio.v1 import revocation_pb2 as revocation_pb2
+from capiscio_sdk._rpc.gen.capiscio.v1 import revocation_pb2_grpc as revocation_pb2_grpc
+from capiscio_sdk._rpc.gen.capiscio.v1 import scoring_pb2 as scoring_pb2
+from capiscio_sdk._rpc.gen.capiscio.v1 import scoring_pb2_grpc as scoring_pb2_grpc
+from capiscio_sdk._rpc.gen.capiscio.v1 import simpleguard_pb2 as simpleguard_pb2
+from capiscio_sdk._rpc.gen.capiscio.v1 import simpleguard_pb2_grpc as simpleguard_pb2_grpc
+from capiscio_sdk._rpc.gen.capiscio.v1 import trust_pb2 as trust_pb2
+from capiscio_sdk._rpc.gen.capiscio.v1 import trust_pb2_grpc as trust_pb2_grpc